User profile put fix

+++ app/src/main/java/com/jder00138218/liftapp/ui/register/viewmodel/RegisterViewModel.kt
@@ -23,7 +23,7 @@ class RegisterViewModel(private val repository: CredentialsRepository) : ViewMod
     private var _passwordVe by mutableStateOf("")
     private var _genre by mutableStateOf("")
     private var _date by mutableStateOf("")
-    private var _weigth by mutableStateOf(0)
+    private var _weigth by mutableStateOf(0.0)
     private var _height by mutableStateOf(0.0)
     private var _isVisiblePaswd by mutableStateOf(false)
     private var _status = mutableStateOf<RegisterUiStatus>(RegisterUiStatus.Resume)
@@ -67,7 +67,7 @@ class RegisterViewModel(private val repository: CredentialsRepository) : ViewMod
             _date = value
         }
 
-    var weigth: Int
+    var weigth: Double
         get() = _weigth
         set(value) {
             _weigth = value
@@ -88,7 +88,7 @@ class RegisterViewModel(private val repository: CredentialsRepository) : ViewMod
         password: String,
         genre: String,
         date: String,
-        weigth: Int,
+        weigth: Double,
         heigth: Double
     ) {
         viewModelScope.launch {
@@ -127,7 +127,7 @@ class RegisterViewModel(private val repository: CredentialsRepository) : ViewMod
             password.isEmpty() -> return false
             genre.isEmpty() -> return false
             date.isEmpty() -> return false
-            (weigth == 0) -> return false
+            (weigth == 0.0) -> return false
             (heigth == 0.0) -> return false
         }
         return true
@@ -145,7 +145,7 @@ class RegisterViewModel(private val repository: CredentialsRepository) : ViewMod
         _genre = ""
         _date = ""
         _height = 0.0
-        _weigth = 0
+        _weigth = 0.0
     }
